+++ src/controllers/MemberController.js
@@ -6,6 +6,8 @@ const path = require('path');
 const jwt = require('jsonwebtoken');
 const authConfig = require('../config/auth');
 
+const errors = []
+
 const { promisify } = require('util');
 
 const generateHash = password => bcrypt.hashSync(password, bcrypt.genSaltSync(8), null);
@@ -16,33 +18,45 @@ const generateToken = (params = {}) => jwt.sign(params, authConfig.secretMember,
 
 module.exports = {
   async index(req, res) {
-    const { jeId } = req.params;
-    try {
-      const je = await Je.findByPk(jeId, {
-        include: { association: 'member' },
-      });
-      if (!je)
-        return res.status(404).json({ msg: 'ENTERPRISE NOT FOUND' })
 
-      if (je.member.length == 0)
-        return res.status(404).json({ msg: 'NO MEMBER FOUND' })
+    const { jeId } = req.params;
+    if (!jeId || jeId == undefined || jeId == null)
+      return res.status(400).json({ error: 'JE ID IS INVALID' })
+
+    try {
+      const je = await Je.findByPk(jeId);
+      if (!je)
+        return res.status(404).json({ error: 'ENTERPRISE NOT FOUND' })
+
+      const members = await Member.findAll({
+        where: { jeId }
+      })
+      if (members.length == 0)
+        return res.status(404).json({ error: 'NO MEMBER FOUND' })
 
       je.password = undefined;
-      let member = je.member;
-      je.dataValues.member = undefined;
-      for (let i = 0; i < member.length; i++)
-        member[i].password = undefined;
+      for (let i = 0; i < members.length; i++)
+        members[i].password = undefined;
 
-      return res.status(200).json({ je, member });
+      console.log(members)
+
+      return res.status(200).json({ je, members });
 
     } catch (error) {
-      return res.status(400).json(error);
+      return res.status(400).json({ error: 'ERROR WHEN GET MEMBER' });
     }
   },
 
-  async store(req, res) {
+  async store(req, res) {  //AQUI
     const { jeId } = req.params;
-    const { email, password, name, board, position, sr, dutyDate, dutyTime } = req.body;
+    if (!jeId || jeId == null || jeId == undefined) errors.push({ error: 'JE ID IS INVALID' })
+    const { email, password, name, board, position, sr, image, dutyDate, dutyTime } = req.body;
+    if (!email || email == null || email == undefined) errors.push({ error: 'EMAIL IS INVALID' })
+    if (!password || password == null || password == undefined) errors.push({ error: 'PASSWORD IS INVALID' })
+    if (!name || name == null || name == undefined) errors.push({ error: 'NAME IS INVALID' })
+    if (!sr || sr == null || sr == undefined) errors.push({ error: 'SR IS INVALID' })
+    if (errors.length > 0) return res.status(400).json(errors)
+
     try {
       const je = await Je.findByPk(jeId);
 
@@ -80,23 +94,34 @@ module.exports = {
 
   async delete(req, res) {
     const { id } = req.body;
+    if (!id || id == null || id == undefined)
+      return res.status(400).json({ error: 'MEMBER ID IS INVALID' })
+
     try {
       const member = await Member.findByPk(id);
       if (member) {
         if (member.image)
           promisify(fs.unlink)(path.resolve(__dirname, '..', '..', 'public', 'uploads', 'member', member.image));
         member.destroy();
-        return res.status(200).json({ msg: 'ok' });
+        return res.status(200).json({ msg: 'MEMBER DELETED SUCCESSFULLY' });
       }
       else
-        return res.status(400).json({ msg: 'NOT FOUND' });
+        return res.status(404).json({ error: 'MEMBER NOT FOUND' });
     } catch (error) {
-      return res.status(400).json(error);
+      return res.status(400).json({ error: 'MEMBER DELETE ERROR' });
     }
   },
 
-  async update(req, res) {
-    const { id, name, board, position, sr, dutyDate, dutyTime } = req.body;
+  async update(req, res) { //AQUI
+    const { id, name, board, password, position, sr, image, dutyDate, dutyTime } = req.body;
+    if (!email || email == null || email == undefined) errors.push({ error: 'EMAIL IS INVALID' })
+    if (!password || password == null || password == undefined) errors.push({ error: 'PASSWORD IS INVALID' })
+    if (!name || name == null || name == undefined) errors.push({ error: 'NAME IS INVALID' })
+    if (!board || board == null || board == undefined) errors.push({ error: 'BOARD IS INVALID' })
+    if (!position || position == null || position == undefined) errors.push({ error: 'POSITION IS INVALID' })
+    if (!sr || sr == null || sr == undefined) errors.push({ error: 'SR IS INVALID' })
+    if (errors.length > 0) return res.status(400).json(errors)
+
     try {
       const member = await Member.findByPk(id);
       if (member) {
